Support runtime IDataObject when built in COM interop turned off
