+++ src/System.Windows.Forms/tests/ComDisabledTests/GlobalUsings.cs
@@ -1,0 +1,6 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+global using FluentAssertions;
+global using Windows.Win32;
+global using Xunit;
